Improve e2e file structure and add new tests
